+++ core/Include/radar_window.h
@@ -21,12 +21,12 @@ void initialize_windowing(WindowType type);
  * @brief 将选定的窗函数应用于输入数据。
  * @param type 选定的窗类型。
  * @param input_data 原始 int8_t 数据。
- * @param output_windowed 加窗后的 double 数据。
+ * @param output_windowed 加窗后的 float 数据。
  */
 void apply_range_windowing(
     WindowType type,
     const int8_t input_data[RADAR_ANT_COUNT][RADAR_CHIRP_COUNT][RADAR_CHIRP_POINTS],
-    double output_windowed[RADAR_ANT_COUNT][RADAR_CHIRP_COUNT][RADAR_CHIRP_POINTS]);
+    float output_windowed[RADAR_ANT_COUNT][RADAR_CHIRP_COUNT][RADAR_CHIRP_POINTS]);
 /**
  * @brief 将选定的窗函数应用于1D FFT后的数据（速度维
  * @param data 1D FFT后的复数数据。
